Refactor: QueryContext is now a struct Breaking: `IQueryContext` -> Just use concrete `QueryContext` Affects `IQueryPart.Evaluate`
